+++ Healthy/Classes/Modules/MainTabBarController/MainTabBarController.swift
@@ -1,9 +1,9 @@
 import UIKit
 
 final class MainTabBarController: UITabBarController {
-    
+
     // MARK: Properties
-    
+
     private let addButton: UIButton = {
         let button = UIButton(type: .system)
         button.translatesAutoresizingMaskIntoConstraints = false
@@ -13,24 +13,51 @@ final class MainTabBarController: UITabBarController {
         button.layer.cornerRadius = 28
         return button
     }()
-    
+
     // MARK: Lifecycle
-    
+
     override func viewDidLoad() {
         super.viewDidLoad()
         configureTabBar()
     }
-    
+
     // MARK: Configurations
-    
+
     private func configureTabBar() {
         configureMainTabBar()
         addCenterButton(addButton, withSize: CGSize(width: 58, height: 58))
         configureViewControllers()
     }
-    
-    // MARK: Configure TabBar ViewControllers
-    
+}
+
+// MARK: Add Custom TabBar and CenterButton
+
+private extension MainTabBarController {
+
+    func configureMainTabBar() {
+        let customTabBar = MainTabBar()
+        setValue(customTabBar, forKey: "tabBar")
+    }
+
+    func addCenterButton(_ button: UIButton, withSize size: CGSize) {
+        tabBar.addSubview(button)
+
+        button.layer.cornerRadius = size.height / 2
+        button.translatesAutoresizingMaskIntoConstraints = false
+
+        NSLayoutConstraint.activate([
+            button.centerXAnchor.constraint(equalTo: tabBar.centerXAnchor),
+            button.topAnchor.constraint(equalTo: tabBar.topAnchor, constant: -22),
+            button.heightAnchor.constraint(equalToConstant: size.height),
+            button.widthAnchor.constraint(equalToConstant: size.width)
+        ])
+    }
+}
+
+// MARK: Configure TabBar ViewControllers
+
+private extension MainTabBarController {
+
     private func configureViewControllers() {
         let homeViewController = makeHomeViewController()
         let notificationViewController = makeNotificationViewController()
@@ -43,7 +70,7 @@ final class MainTabBarController: UITabBarController {
             unionViewController
         ]
     }
-    
+
     private func makeHomeViewController() -> UIViewController {
         let viewController = UIViewController()
         viewController.view.backgroundColor = .red
@@ -54,7 +81,7 @@ final class MainTabBarController: UITabBarController {
         )
         return viewController
     }
-    
+
     private func makeNotificationViewController() -> UIViewController {
         let viewController = UIViewController()
         viewController.view.backgroundColor = .yellow
@@ -65,7 +92,7 @@ final class MainTabBarController: UITabBarController {
         )
         return viewController
     }
-    
+
     private func makeProfileViewController() -> UIViewController {
         let viewController = UIViewController()
         viewController.view.backgroundColor = .systemBrown
@@ -76,7 +103,7 @@ final class MainTabBarController: UITabBarController {
         )
         return viewController
     }
-    
+
     private func makeUnionViewController() -> UIViewController {
         let viewController = UIViewController()
         viewController.view.backgroundColor = .systemPink
@@ -88,5 +115,3 @@ final class MainTabBarController: UITabBarController {
         return viewController
     }
 }
-
-
